+++ packages/medical-domain/src/domains/models/medical/datum/bolus.model.ts
@@ -29,6 +29,7 @@ import type BaseDatum from './basics/base-datum.model'
 import type Wizard from './wizard.model'
 import { BolusSubtype } from './enums/bolus-subtype.enum'
 import { type DatumType } from './enums/datum-type.enum'
+import Prescriptor from './enums/prescriptor.enum'
 
 function isBolusSubType(value: unknown): boolean {
   return Object.values(BolusSubtype).includes(value as BolusSubtype)
@@ -39,7 +40,7 @@ type Bolus = BaseDatum & {
   subType: BolusSubtype
   uploadId: string
   normal: number
-  prescriptor: string
+  prescriptor: Prescriptor
   wizard: Wizard | null
   expectedNormal?: number
   insulinOnBoard?: number