+++ packages/medical-domain/src/domains/models/statistics/carbs-statistics.model.ts
@@ -25,6 +25,8 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+import { HoursRange } from './satistics.model'
+
 export interface CarbsStatistics {
   totalCarbsPerDay: number
   rescueCarbsPerDay: number
@@ -33,3 +35,11 @@ export interface CarbsStatistics {
   totalMealCarbsWithRescueCarbsEntries: number
   totalRescueCarbsEntries: number
 }
+
+export interface RescueCarbsAveragePerRange {
+  numberOfIntakes: number
+  confirmedCarbs: number
+  recommendedCarbs: number
+}
+
+export type RescueCarbsAverageStatistics = Map<HoursRange, RescueCarbsAveragePerRange>