+++ packages/medical-domain/test/data-generator.ts
@@ -27,7 +27,6 @@
 
 import { faker } from '@faker-js/faker'
 import type Bolus from '../src/domains/models/medical/datum/bolus.model'
-import { BolusSubtype } from '../src/domains/models/medical/datum/enums/bolus-subtype.enum'
 import type Basal from '../src/domains/models/medical/datum/basal.model'
 import type Cbg from '../src/domains/models/medical/datum/cbg.model'
 import { bgUnits } from '../src/domains/models/medical/datum/bg.model'
@@ -46,7 +45,7 @@ import type WarmUp from '../src/domains/models/medical/datum/warm-up.model'
 import type Wizard from '../src/domains/models/medical/datum/wizard.model'
 import type ZenMode from '../src/domains/models/medical/datum/zen-mode.model'
 import type Datum from '../src/domains/models/medical/datum.model'
-import { DatumType } from '../src/domains/models/medical/datum/enums/datum-type.enum'
+import { DatumType, Prescriptor, BolusSubtype } from '../src'
 import Source from '../src/domains/models/medical/datum/enums/source.enum'
 import DurationUnit from '../src/domains/models/medical/datum/enums/duration-unit.enum'
 import Unit from '../src/domains/models/medical/datum/enums/unit.enum'
@@ -111,7 +110,7 @@ function createRandomBolus(date?: Date): Bolus {
     subType: faker.helpers.arrayElement(Object.values(BolusSubtype)),
     uploadId: faker.string.uuid(),
     normal: 0,
-    prescriptor: 'test',
+    prescriptor: Prescriptor.Auto,
     wizard: null
   }
 }