+++ packages/yourloops/components/carbs-and-bolus/manual-bolus-tooltip.tsx
@@ -25,40 +25,39 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import {
-  checkBannerLanguageChange,
-  checkCaregiverHeader,
-  checkHcpHeader,
-  checkPatientHeader,
-  type HeaderInfo
-} from '../assert/header.assert'
-import { checkFooterForCaregiver, checkFooterForHcp, checkFooterForPatient } from '../assert/footer.assert'
+import React, { FC } from 'react'
+import Typography from '@mui/material/Typography'
+import Box from '@mui/material/Box'
+import { MANUAL_BOLUS_COLOR, useCarbsAndBolusStyles } from './carbs-and-bolus-styles'
+import { useTranslation } from 'react-i18next'
+import { ManualBolusAveragePerRange } from 'medical-domain'
 
-export interface AppMainLayoutHcpParams {
-  footerHasLanguageSelector?: boolean
-  headerInfo: HeaderInfo
-}
+export const ManualBolusTooltip: FC<{ manualBolus: ManualBolusAveragePerRange }> = ({ manualBolus }) => {
+  const { classes } = useCarbsAndBolusStyles()
+  const { t } = useTranslation()
 
-export interface AppMainLayoutParams {
-  footerHasLanguageSelector?: boolean
-  loggedInUserFullName: string
-}
-
-export const testAppMainLayoutForHcp = async (appMainLayoutParams: AppMainLayoutHcpParams) => {
-  await checkHcpHeader(appMainLayoutParams.headerInfo)
-  checkFooterForHcp(appMainLayoutParams.footerHasLanguageSelector ?? false)
-}
-
-export const testAppMainLayoutForCaregiver = async (appMainLayoutParams: AppMainLayoutParams) => {
-  await checkCaregiverHeader(appMainLayoutParams.loggedInUserFullName)
-  checkFooterForCaregiver(appMainLayoutParams.footerHasLanguageSelector ?? false)
-}
-
-export const testAppMainLayoutForPatient = async (appMainLayoutParams: AppMainLayoutParams) => {
-  await checkPatientHeader(appMainLayoutParams.loggedInUserFullName)
-  checkFooterForPatient(appMainLayoutParams.footerHasLanguageSelector ?? false)
-}
-
-export const testBannerLanguageUpdate = async () => {
-  await checkBannerLanguageChange()
+  return (
+    <Box
+      className={classes.hoverTooltip}
+      sx={{ border: `2px solid ${MANUAL_BOLUS_COLOR}` }}
+    >
+      <div className={`${classes.tooltipTail} manual-bolus`} />
+      <Typography
+        variant="subtitle2"
+        className="header"
+      >
+        {t('manual-bolus')}
+      </Typography>
+      <Box className="content">
+        <div className="flex-justify-between-align-center">
+          <Typography variant="body2">{t('avg-number-of-injections')}</Typography>
+          <Typography variant="body2">{manualBolus.numberOfInjections}</Typography>
+        </div>
+        <div className="flex-justify-between-align-center">
+          <Typography variant="body2">{t('avg-confirmed-dose')}</Typography>
+          <Typography variant="body2">{manualBolus.confirmedDose}U</Typography>
+        </div>
+      </Box>
+    </Box>
+  )
 }