+++ packages/yourloops/components/carbs-and-bolus/models/carbs-and-bolus.model.ts
@@ -25,40 +25,13 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import {
-  checkBannerLanguageChange,
-  checkCaregiverHeader,
-  checkHcpHeader,
-  checkPatientHeader,
-  type HeaderInfo
-} from '../assert/header.assert'
-import { checkFooterForCaregiver, checkFooterForHcp, checkFooterForPatient } from '../assert/footer.assert'
-
-export interface AppMainLayoutHcpParams {
-  footerHasLanguageSelector?: boolean
-  headerInfo: HeaderInfo
-}
-
-export interface AppMainLayoutParams {
-  footerHasLanguageSelector?: boolean
-  loggedInUserFullName: string
-}
-
-export const testAppMainLayoutForHcp = async (appMainLayoutParams: AppMainLayoutHcpParams) => {
-  await checkHcpHeader(appMainLayoutParams.headerInfo)
-  checkFooterForHcp(appMainLayoutParams.footerHasLanguageSelector ?? false)
-}
-
-export const testAppMainLayoutForCaregiver = async (appMainLayoutParams: AppMainLayoutParams) => {
-  await checkCaregiverHeader(appMainLayoutParams.loggedInUserFullName)
-  checkFooterForCaregiver(appMainLayoutParams.footerHasLanguageSelector ?? false)
-}
-
-export const testAppMainLayoutForPatient = async (appMainLayoutParams: AppMainLayoutParams) => {
-  await checkPatientHeader(appMainLayoutParams.loggedInUserFullName)
-  checkFooterForPatient(appMainLayoutParams.footerHasLanguageSelector ?? false)
-}
-
-export const testBannerLanguageUpdate = async () => {
-  await checkBannerLanguageChange()
+export enum CarbsAndBolusTimeRange {
+  Midnight = '0h',
+  Three = '3h',
+  Six = '6h',
+  Nine = '9h',
+  Twelve = '12h',
+  Fifteen = '15h',
+  Eighteen = '18h',
+  TwentyOne = '21h'
 }